Some cleanup processes were separated.
