+++ core-domain/src/main/res/values/strings.xml
@@ -29,6 +29,8 @@
     <string name="sanitizer_change_name" translatable="false">Change</string>
     <string name="sanitizer_cx_name" translatable="false">CX Analytics</string>
     <string name="sanitizer_douyin_name">Douyin</string>
+    <string name="sanitizer_dingtalk_name">DingTalk</string>
+    <string name="sanitizer_dianping_name">Dianping</string>
     <string name="sanitizer_ebay_name" translatable="false">eBay</string>
     <string name="sanitizer_echobox_name" translatable="false">Echobox</string>
     <string name="sanitizer_elfinanciero_name" translatable="false">ElFinanciero</string>
@@ -37,7 +39,7 @@
     <string name="sanitizer_facebook_name" translatable="false">Facebook.com</string>
     <string name="sanitizer_fastcompany" translatable="false">Fast Company</string>
     <string name="sanitizer_flipkart_name" translatable="false">Flipkart</string>
-    <string name="sanitizer_feishu_dingtalk_name">Feishu / DingTalk</string>
+    <string name="sanitizer_feishu_name">Feishu</string>
     <string name="sanitizer_adobe_marketo_engage_name" translatable="false">Adobe Marketo Engage</string>
     <string name="sanitizer_ga_name" translatable="false">Google Analytics</string>
     <string name="sanitizer_georiot_name" translatable="false">GeoRiot</string>
@@ -61,7 +63,7 @@
     <string name="sanitizer_meta_ad_name" translatable="false">Meta Ads</string>
     <string name="sanitizer_mydealz_parameters_name">MyDealz Parameters</string>
     <string name="sanitizer_mydealz_redirects_name">MyDealz Redirects</string>
-    <string name="sanitizer_meituan_dianping_name">Meituan / Dianping</string>
+    <string name="sanitizer_meituan_name">Meituan</string>
     <string name="sanitizer_netflix_name" translatable="false">Netflix</string>
     <string name="sanitizer_newegg_name" translatable="false">NewEgg</string>
     <string name="sanitizer_pearl" translatable="false">Pearl</string>
